refactor: split export.rs into focused modules (types, stats, snapshot, timestamp) (#980) (#991)
* refactor: split export.rs into focused modules (types, stats, snapshot, timestamp) (#980)

Split the monolithic src/export.rs (778 lines) into a src/export/ module
directory with four focused sub-modules. All public API re-exported from
mod.rs for backward compatibility. No functional changes.


* chore: auto-increment versions for changed projects

* fix: use correct serde field names in GPU error test (#991)

The test JSON used camelCase field names (`neuronType`, `fromUuid`,
`toUuid`) but NeuronJson uses `#[serde(rename = "type")]` and
SynapseJson uses snake_case (`from_uuid`/`to_uuid`). On GPU-equipped
machines the test skipped early so this was never caught; on CI (no GPU)
the deserialisation failed producing a `data_validation` error instead
of reaching the GPU availability check.


---------
