+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "neat_ai_discovery"
-version = "0.72.14"
+version = "0.72.15"
 edition = "2024"
 license = "Apache-2.0"
