+++ src/export/snapshot.rs
@@ -1,17 +1,16 @@
-//! Visualisation snapshot export module
+//! Main visualisation snapshot export pipeline.
 //!
-//! Exports a debug-friendly JSON snapshot of a creature and its recorded
-//! activations/errors for use with the NEAT-AI-Explore visualiser.
-//!
-//! This is an optional debug tool that does not affect existing analysis behaviour.
+//! Reads Parquet recordings, builds per-neuron and per-synapse columnar data,
+//! computes impact scores and reconstruction checks, then writes the full
+//! snapshot to a JSON file.
 
 #![allow(
     clippy::cast_possible_truncation,
     clippy::cast_possible_wrap,
     clippy::cast_precision_loss
 )] // Intentional numeric casts for GPU/neural network computation (Issue #873)
+
 use anyhow::{Context, Result};
-use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
 use std::fs::File;
 use std::io::BufWriter;
@@ -21,265 +20,9 @@ use crate::focus::compute_impacts_public;
 use crate::parquet_format::read_records_from_parquet_with_limit;
 use crate::types::DiscoverRecord;
 
-/// Snapshot metadata
-#[derive(Debug, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct SnapshotMeta {
-    pub exported_at: String,
-    pub discovery_version: String,
-    pub parquet_file: String,
-    #[serde(skip_serializing_if = "Option::is_none", default)]
-    pub notes: Option<String>,
-}
-
-/// Per-neuron stats (aggregates)
-#[derive(Debug, Serialize, Deserialize, Default)]
-#[serde(rename_all = "camelCase")]
-pub struct NeuronRecordingStats {
-    pub mean_activation: f32,
-    pub activation_variance: f32,
-    pub activation_min: f32,
-    pub activation_max: f32,
-    /// Arithmetic mean of errors (can be misleading as +/- cancel out)
-    pub mean_error: f32,
-    /// Mean Absolute Error - used in focus neuron ranking
-    pub mean_absolute_error: f32,
-    /// Mean Squared Error - matches production creature scoring
-    pub mean_squared_error: f32,
-    pub error_variance: f32,
-    pub error_min: f32,
-    pub error_max: f32,
-    pub record_count: usize,
-}
-
-/// Per-neuron recorded data (columnar format)
-#[derive(Debug, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct NeuronRecording {
-    /// Activation values per obsIndex (aligned with obsIndices array)
-    pub activation: Vec<f32>,
-    /// Value (pre-activation) per obsIndex; null encoded as JSON null
-    pub value: Vec<Option<f32>>,
-    /// Errors per obsIndex (each element is a vector of error values)
-    pub errors: Vec<Vec<f32>>,
-    /// Aggregated stats
-    pub stats: NeuronRecordingStats,
-}
-
-/// Per-synapse stats (aggregates)
-#[derive(Debug, Serialize, Deserialize, Default)]
-#[serde(rename_all = "camelCase")]
-pub struct SynapseStats {
-    pub mean_contribution: f32,
-    pub contribution_variance: f32,
-    pub contribution_min: f32,
-    pub contribution_max: f32,
-}
-
-/// Per-synapse derived data (columnar format)
-#[derive(Debug, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct SynapseDerived {
-    pub from_uuid: String,
-    pub to_uuid: String,
-    pub weight: f32,
-    /// Contribution = fromActivation * weight per obsIndex
-    pub contribution: Vec<f32>,
-    /// Aggregated stats
-    pub stats: SynapseStats,
-}
-
-/// Reconstruction check results for a single neuron
-#[derive(Debug, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct ReconstructionCheck {
-    pub neuron_uuid: String,
-    pub squash: String,
-    pub bias: f32,
-    /// max|recordedValue - reconstructedValue|
-    pub max_value_delta: f32,
-    /// max|recordedActivation - reconstructedActivation|
-    pub max_activation_delta: f32,
-    /// mean|valueDelta|
-    pub mean_value_delta: f32,
-    /// mean|activationDelta|
-    pub mean_activation_delta: f32,
-    /// Top-K worst obsIndex + deltas
-    pub worst_samples: Vec<ReconstructionSample>,
-}
-
-/// A single sample's reconstruction delta
-#[derive(Debug, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct ReconstructionSample {
-    pub obs_index: u32,
-    pub recorded_value: Option<f32>,
-    pub reconstructed_value: f32,
-    pub value_delta: f32,
-    pub recorded_activation: f32,
-    pub reconstructed_activation: f32,
-    pub activation_delta: f32,
-}
-
-/// Derived data section
-#[derive(Debug, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct DerivedData {
-    /// Impact scores keyed by neuron UUID
-    pub impacts_by_neuron_uuid: HashMap<String, f32>,
-    /// Synapse-level derived data keyed by "fromUuid->toUuid"
-    pub synapses: HashMap<String, SynapseDerived>,
-    /// Reconstruction checks per neuron (non-input neurons only)
-    #[serde(skip_serializing_if = "Option::is_none", default)]
-    pub reconstruction_checks: Option<Vec<ReconstructionCheck>>,
-}
-
-/// Recording data section
-#[derive(Debug, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct RecordingData {
-    /// Ordered obsIndex values
-    pub obs_indices: Vec<u32>,
-    /// Per-neuron recorded data keyed by neuron UUID
-    pub neurons: HashMap<String, NeuronRecording>,
-}
-
-/// The full snapshot structure written to JSON
-#[derive(Debug, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct VisualisationSnapshot {
-    pub meta: SnapshotMeta,
-    pub creature: CreatureJson,
-    pub recording: RecordingData,
-    pub derived: DerivedData,
-}
-
-/// Export options
-pub struct ExportOptions {
-    pub include_per_synapse_series: bool,
-    pub include_reconstruction_checks: bool,
-    pub max_obs: Option<u32>,
-    pub top_k_worst_samples: usize,
-}
-
-impl Default for ExportOptions {
-    fn default() -> Self {
-        Self {
-            include_per_synapse_series: true,
-            include_reconstruction_checks: true,
-            max_obs: None,
-            top_k_worst_samples: 20,
-        }
-    }
-}
-
-/// Export result statistics
-#[derive(Debug, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct ExportStats {
-    pub obs_count: usize,
-    pub neuron_count: usize,
-    pub synapse_count: usize,
-    pub output_count: usize,
-}
-
-/// Apply a squash function by name.
-///
-/// # Performance (Issue #211)
-/// Delegates to `crate::activations::apply_scalar_squash` which uses `normalise_squash_name`
-/// with a fast path for already-uppercase strings (zero allocation in that case).
-fn apply_squash(squash: &str, value: f32) -> f32 {
-    // Use the centralised activation function implementation.
-    // Returns None for aggregate squashes (MINIMUM/MAXIMUM/IF/etc.) which cannot be
-    // represented as f(value). For those, default to identity (passthrough).
-    crate::activations::apply_scalar_squash(squash, value).unwrap_or(value)
-}
-
-/// Compute stats from a slice of f32 values
-fn compute_stats(values: &[f32]) -> (f32, f32, f32, f32) {
-    if values.is_empty() {
-        return (0.0, 0.0, 0.0, 0.0);
-    }
-
-    // We intentionally ignore non-finite values (NaN/±Infinity) so they cannot
-    // corrupt summary statistics or JSON serialisation.
-    //
-    // Important: Even with finite inputs, naive f32 accumulation can overflow
-    // internally (eg variance for `[0.0, f32::MAX]`). We compute in f64 and then
-    // clamp to JSON-safe f32 outputs.
-    //
-    // When there are no finite values, return zeros rather than ±Infinity.
-    // This matches the empty-slice behaviour and keeps the output JSON valid.
-    let mut count: u64 = 0;
-    let mut mean: f64 = 0.0;
-    let mut m2: f64 = 0.0;
-    let mut min: f64 = 0.0;
-    let mut max: f64 = 0.0;
-
-    for &x in values {
-        if !x.is_finite() {
-            continue;
-        }
-
-        let xf = x as f64;
-
-        if count == 0 {
-            count = 1;
-            mean = xf;
-            m2 = 0.0;
-            min = xf;
-            max = xf;
-            continue;
-        }
-
-        if xf < min {
-            min = xf;
-        }
-        if xf > max {
-            max = xf;
-        }
-
-        count += 1;
-        let delta = xf - mean;
-        mean += delta / count as f64;
-        let delta2 = xf - mean;
-        m2 += delta * delta2;
-    }
-
-    if count == 0 {
-        return (0.0, 0.0, 0.0, 0.0);
-    }
-
-    let variance = m2 / count as f64;
-
-    (
-        json_safe_f32(mean as f32),
-        json_safe_f32(variance as f32),
-        json_safe_f32(min as f32),
-        json_safe_f32(max as f32),
-    )
-}
-
-/// Convert a float to a JSON-safe finite value.
-///
-/// `serde_json` will serialise non-finite floats (NaN/±Infinity) as `null`, which
-/// breaks consumers that expect numeric arrays (and breaks round-tripping into
-/// `Vec<f32>`).
-///
-/// We clamp ±Infinity to ±`f32::MAX` and map NaN to 0.0. This keeps exported
-/// snapshots valid JSON while preserving sign/magnitude semantics as much as
-/// possible for debugging.
-fn json_safe_f32(value: f32) -> f32 {
-    if value.is_finite() {
-        value
-    } else if value.is_nan() {
-        0.0
-    } else if value.is_sign_positive() {
-        f32::MAX
-    } else {
-        -f32::MAX
-    }
-}
+use super::stats::{apply_squash, compute_stats, json_safe_f32};
+use super::timestamp::chrono_lite_now;
+use super::types::*;
 
 /// Export a visualisation snapshot to JSON
 pub fn export_visualisation_snapshot(
@@ -631,148 +374,4 @@ pub fn export_visualisation_snapshot(
         .with_context(|| format!("Failed to write JSON to: {out_file}"))?;
 
     Ok(stats)
-}
-
-/// Simple ISO 8601 timestamp without external chrono dependency
-fn chrono_lite_now() -> String {
-    use std::time::{SystemTime, UNIX_EPOCH};
-
-    let duration = SystemTime::now()
-        .duration_since(UNIX_EPOCH)
-        .unwrap_or_default();
-    let secs = duration.as_secs();
-
-    // Convert to date/time components (UTC)
-    let days = secs / 86400;
-    let time_secs = secs % 86400;
-    let hours = time_secs / 3600;
-    let minutes = (time_secs % 3600) / 60;
-    let seconds = time_secs % 60;
-
-    // Simple year/month/day calculation (good enough for logging)
-    let mut year = 1970;
-    let mut remaining_days = days as i64;
-
-    loop {
-        let days_in_year = if is_leap_year(year) { 366 } else { 365 };
-        if remaining_days < days_in_year {
-            break;
-        }
-        remaining_days -= days_in_year;
-        year += 1;
-    }
-
-    let days_in_months: [i64; 12] = if is_leap_year(year) {
-        [31, 29, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31]
-    } else {
-        [31, 28, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31]
-    };
-
-    let mut month = 1;
-    for &days_in_month in &days_in_months {
-        if remaining_days < days_in_month {
-            break;
-        }
-        remaining_days -= days_in_month;
-        month += 1;
-    }
-    let day = remaining_days + 1;
-
-    format!("{year:04}{month:02}{day:02}T{hours:02}{minutes:02}{seconds:02}Z")
-}
-
-fn is_leap_year(year: i64) -> bool {
-    (year % 4 == 0 && year % 100 != 0) || (year % 400 == 0)
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_apply_squash_identity() {
-        assert!((apply_squash("IDENTITY", 0.5) - 0.5).abs() < 1e-6);
-        assert!((apply_squash("IDENTITY", -1.0) - (-1.0)).abs() < 1e-6);
-    }
-
-    #[test]
-    fn test_apply_squash_tanh() {
-        assert!((apply_squash("TANH", 0.0) - 0.0).abs() < 1e-6);
-        assert!((apply_squash("TANH", 1.0) - 1.0_f32.tanh()).abs() < 1e-6);
-    }
-
-    #[test]
-    fn test_apply_squash_relu() {
-        assert!((apply_squash("RELU", 1.0) - 1.0).abs() < 1e-6);
-        assert!((apply_squash("RELU", -1.0) - 0.0).abs() < 1e-6);
-    }
-
-    #[test]
-    fn test_apply_squash_hard_tanh() {
-        assert!((apply_squash("HARD_TANH", 0.5) - 0.5).abs() < 1e-6);
-        assert!((apply_squash("HARD_TANH", 2.0) - 1.0).abs() < 1e-6);
-        assert!((apply_squash("HARD_TANH", -2.0) - (-1.0)).abs() < 1e-6);
-    }
-
-    #[test]
-    fn test_compute_stats_empty() {
-        let (mean, var, min, max) = compute_stats(&[]);
-        assert_eq!(mean, 0.0);
-        assert_eq!(var, 0.0);
-        assert_eq!(min, 0.0);
-        assert_eq!(max, 0.0);
-    }
-
-    #[test]
-    fn test_compute_stats_single() {
-        let (mean, var, min, max) = compute_stats(&[5.0]);
-        assert!((mean - 5.0).abs() < 1e-6);
-        assert!((var - 0.0).abs() < 1e-6);
-        assert!((min - 5.0).abs() < 1e-6);
-        assert!((max - 5.0).abs() < 1e-6);
-    }
-
-    #[test]
-    fn test_compute_stats_multiple() {
-        let values = vec![1.0, 2.0, 3.0, 4.0, 5.0];
-        let (mean, var, min, max) = compute_stats(&values);
-        assert!((mean - 3.0).abs() < 1e-6);
-        assert!((var - 2.0).abs() < 1e-6); // variance of [1,2,3,4,5] = 2
-        assert!((min - 1.0).abs() < 1e-6);
-        assert!((max - 5.0).abs() < 1e-6);
-    }
-
-    #[test]
-    fn test_compute_stats_ignores_non_finite_values() {
-        // Mean/variance should be computed over finite values only.
-        // This avoids skewing the result when recordings contain NaN/±Infinity.
-        let values = [1.0_f32, 2.0, f32::NAN, 3.0];
-        let (mean, var, min, max) = compute_stats(&values);
-
-        assert!((mean - 2.0).abs() < 1e-6);
-        assert!((var - (2.0 / 3.0)).abs() < 1e-6);
-        assert!((min - 1.0).abs() < 1e-6);
-        assert!((max - 3.0).abs() < 1e-6);
-    }
-
-    #[test]
-    fn test_compute_stats_all_non_finite_returns_zeros() {
-        // Returning ±Infinity here is semantically wrong and can break JSON output.
-        let values = [f32::NAN, f32::INFINITY, f32::NEG_INFINITY];
-        let (mean, var, min, max) = compute_stats(&values);
-
-        assert_eq!(mean, 0.0);
-        assert_eq!(var, 0.0);
-        assert_eq!(min, 0.0);
-        assert_eq!(max, 0.0);
-    }
-
-    #[test]
-    fn test_chrono_lite_now_format() {
-        let timestamp = chrono_lite_now();
-        // Should be in format YYYYMMDDTHHMMSSZ
-        assert_eq!(timestamp.len(), 16);
-        assert!(timestamp.ends_with('Z'));
-        assert!(timestamp.contains('T'));
-    }
 }
